requirements.txt
Pillow
keras
tensorflow
sklearn
image_slicer
pydot
matplotlib
